[Debug] Organize pass instrumentation infrastructure
