+++ tilelang/backend/device_codegen.py
@@ -9,7 +9,7 @@ from tvm import IRModule
 from tvm.target import Target
 
 from tilelang import tvm
-from tilelang.utils.pass_events import run_codegen_with_instrumentation
+from tilelang.instrumentation import run_codegen_with_instrumentation
 
 DeviceCodegenFunc = Callable[[IRModule, Target], IRModule]
