+++ tilelang/instrumentation/__init__.py
@@ -1,0 +1,51 @@
+"""Shared compiler instrumentation infrastructure for TileLang tools."""
+
+from .events import (
+    CapturePredicate,
+    IncompletePass,
+    PassEvent,
+    PassEventObserver,
+    SequenceAllocator,
+    StackedPassInstrument,
+    active_stacked_pass_instruments,
+    current_pass_phase,
+    pass_phase,
+)
+from .session import (
+    CodegenEvent,
+    CompilePassInstrumentationSession,
+    PassInstrumentationTool,
+    compile_pass_instrumentation,
+    create_pass_instruments,
+    current_compile_pass_instrumentation,
+    current_pass_instrument_context,
+    instrument_current_pass_context,
+    pass_pipeline,
+    register_pass_instrumentation_tool,
+    run_codegen_with_instrumentation,
+    unregister_pass_instrumentation_tool,
+)
+
+__all__ = [
+    "CapturePredicate",
+    "CodegenEvent",
+    "CompilePassInstrumentationSession",
+    "IncompletePass",
+    "PassEvent",
+    "PassEventObserver",
+    "PassInstrumentationTool",
+    "SequenceAllocator",
+    "StackedPassInstrument",
+    "active_stacked_pass_instruments",
+    "compile_pass_instrumentation",
+    "create_pass_instruments",
+    "current_compile_pass_instrumentation",
+    "current_pass_instrument_context",
+    "current_pass_phase",
+    "instrument_current_pass_context",
+    "pass_phase",
+    "pass_pipeline",
+    "register_pass_instrumentation_tool",
+    "run_codegen_with_instrumentation",
+    "unregister_pass_instrumentation_tool",
+]
