+++ tilelang/utils/__init__.py
@@ -20,4 +20,3 @@ from .language import (
 )
 from .deprecated import deprecated  # noqa: F401
 from .version import build_date  # noqa: F401
-from .pass_timing import PassTimingRecord, PassTimingTool, TileLangPassTimingInstrument  # noqa: F401
