--- tilelang/utils/pass_timing.py
@@ -1,19 +1,0 @@
-"""Compatibility exports for the pass-timing developer tool.
-
-The implementation lives in :mod:`tilelang.tools.pass_timing`.  This module
-preserves the historical import path for downstream users.
-"""
-
-from tilelang.tools.pass_timing import (
-    PassTimingRecord,
-    PassTimingTool,
-    TileLangPassTimingInstrument,
-    create_pass_timing_tool,
-)
-
-__all__ = [
-    "PassTimingRecord",
-    "PassTimingTool",
-    "TileLangPassTimingInstrument",
-    "create_pass_timing_tool",
-]
